server1/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7; python_version >= '3'
click==8.0.3; python_version >= '3.6'
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0; python_version >= '3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==0.5.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==21.1
django-heroku==0.0.0
django-seed==0.3.1
django-templated-mail==1.1.1
django==3.2.9
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
djoser==2.1.0
drf-nested-routers==0.93.4
drf-writable-nested==0.6.3
drf-yasg==1.20.0
faker==9.8.3; python_version >= '3.6'
gunicorn==20.1.0
h11==0.12.0; python_version >= '3.6'
httptools==0.3.0
idna==3.3; python_version >= '3'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
mysqlclient==2.1.0
oauthlib==3.1.1; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
psycopg2-binary==2.9.2
pycparser==2.21
pyjwt==2.3.0; python_version >= '3.6'
pyparsing==3.0.6; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ruamel.yaml.clib==0.2.6; python_version < '3.10' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.17; python_version >= '3'
setuptools==59.2.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==4.0.0
social-auth-core==4.1.0; python_version >= '3.6'
sqlparse==0.4.2; python_version >= '3.5'
text-unidecode==1.3
toposort==1.7
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_full_version < '4.0.0'
uvicorn==0.15.0
uvloop==0.16.0
whitenoise==5.3.0
